Rework parsing of the namespace argument
In Python 3, we cannot accept (easily) bytes or None, so we have to
change how we accept the namespace argument.

Previously, if the argument was not passed or it was None, it was
considered missing. Since the None option is no longer possible, we
change it so that if not passed or passed as an empty string, it is
considered missing. This changes the behaviour somewhat, but I hope
that empty namespaces are not used (that's what I understand from
reading various pages on the internet).